PLT-294 Added unit tests to require version
They are failing of course. So I guess I have to fix that.
