+++ src/require_version.h
@@ -1,0 +1,26 @@
+#pragma once
+
+#ifdef __cplusplus
+extern "C"
+{
+#endif
+
+#include <ctype.h>
+#include <string.h>
+#include <stdbool.h>
+
+
+typedef struct {
+    int major;
+    int minor;
+    int patch;
+} version_t;
+
+
+bool parse_version(const char *version_str, version_t *version_out);
+
+bool compare_version(const version_t *version, const char *constraint);
+
+#ifdef __cplusplus
+}
+#endif
